[op][n_upsable][dynamic_sized_window] Implement dynamic sized window
:Release Notes:
Implement dynamic sized window.
Upscale video to fullscreen providing additional contents scale to media
stack. Keep size of size-overriding apps when relaunched.

:Detailed Notes:
If the application's actual resolution is different from platform resolution,
then WAM & webOS Webview adjusts app's content size to fit the wayland window.
It, however, is redundant to create a larger window than app's resolution,
in terms of WAM's memory usage.

Moreover, additional cost may be needed to scale content up.
This patch makes wayland window have app's actual resolution.
In webOS, system compositor scales window to match platform
screen size when window is displayed on screen.

Playing video with webOS media stack it is needed to provide the target position for
the video in UI resolution (and indicate if it is fullscreen or not). For
that, the algorithm uses additional contents scale, to adapt the media
resolution to the UI resolution.
This is needed for properly showing fullscreen videos on videos that
override the application resolution at fullscreen (regular way to see apps
in webOS). The window will get the overridden resolution, but it is
upscaled to show fullscreen. Videos need to adapt to the expected
resolution.

Sometimes on relaunching, the wayland window (and
app size) was resized to display screen size, after fullscreen.
There were two problems. One was that on setting the window fullscreen,
the bounds were set to the size of the display. This has been fixed
implementing the ToggleFullscreenWithSize method.

The other problem was that, once application got fullscreen status,
window manager was sending a configure event with the display size.
Solution is - ignoring this. For that this time
we use the WidgetDelegate::CanResize(), that is set
to false in app runtime already. Additionally this patch adds the
handling of the resize behavior that is populated using CanResize
value.

:Testing Performed:
Launch the Viewster app or other 1280x720 app on a 1920x1080 display
and check the user interface is properly shown.

:QA Notes:
Test-cases: GT-20753, GT-15269, GT-27617, GT-23959, GT-27618

:Issues Addressed:
[PLAT-132028] Port : dynamic_sized_window

Change-Id: I92be7636383c0a80d00e20d99a6c67d199509e50
Originally-Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium84/+/296069
Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium87/+/305183
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Joonhyung Hwang <joonhyung.hwang@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Joonhyung Hwang <joonhyung.hwang@lge.com>
Tested-by: Test BTS <test.bts@lge.com>
Reviewed-on: http://gpro.lge.com/c/webosose/chromium87/+/308660
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>