[op][n_upsable][v8][log] Add the log for fatal error of v8 in ReportFatalErrorInMainThread
:Release Notes:
Add the log for fatal error of v8 in ReportFatalErrorInMainThread to
send the information of fatal error from V8 in release mode.

:Detailed Notes:
DVLOG is only for debug mode, but we need to get the information of
fatal error log from V8 in release mode.
This patch will send the information of fatal error from V8
in release mode.

:Testing Performed:
Checked the log is printed
when ReportFatalErrorInMainThread is forcibly called.

:QA Notes:
Test-cases: N/A (The function is called on the edge case of crash
occurred on critically low memory situation.)

:Issues Addressed:
[PLAT-132013] Port : [op][n_upsable][v8][log] Add the log for fatal error of v8 in ReportFatalErrorInMainThread

Change-Id: I554a792d81b657ce5b6ae2b856b198fc3b0ea042
Originally-Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium84/+/294506
Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium87/+/303126
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Joonhyung Hwang <joonhyung.hwang@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Joonhyung Hwang <joonhyung.hwang@lge.com>
Reviewed-on: http://gpro.lge.com/c/webosose/chromium87/+/308343
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>