[op][n_upsable][ssl] Allow to use an additional certificates database.
:Release Notes:
Allow to use an additional NSS certificates database for self signed
certificates.

:Detailed Notes:
In webOS a system-wide NSS read only database is used for certificates.
But in certain products, it is interesting to be able to maintain an
additional database for storing self-signed certificates added by the user.

This patch provides a new command line parameter for adding an NSS
database at an specified path: --neva-certificates-path. When passed, Chromium
will also load it (as an NSS user certificate database).

:Testing Performed:
1. Create certificate database path: mkdir /tmp/TestCertDB and set
permissions for WAM.
2. Create a certificates database: certutil -N -d /tmp/TestCertDB
3. Add a badssl certificate to the database: certutil -A -d
/tmp/TestCertDB -n badssl_crt_cer -t "pC,," -i badssl.crt.cer
4. Change the permission of files under /tmp/TestCertDB
chmod 777 /tmp/TestCertDB/*
5. Modify WAM to launch with --neva-certificates-path=/tmp/TestCertDB
and restart it.
6. Sideload https://self-signed.badssl.com -> It should load instead of
showing a certificate error.

:QA Notes:
Test-cases: GT-29612

:Issues Addressed:
[PLAT-132020] Port : [op][n_upsable][ssl] Allow to use an additional certificates database.

Change-Id: I9ff3cd0232ad30d353a5c1d2a6526b9c973a172f
Originally-Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium84/+/295872
Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium87/+/303193
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-on: http://gpro.lge.com/c/webosose/chromium87/+/308348
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>