[op][n_upsable][build] Add configurations to build qemux86
:Release Notes:
Add configurations to build qemux86

:Detailed Notes:
Until Chromium72, when building qemux86 we used gcc as a host machine
compiler(instead the clang as default) to fix the V8 compile error.
But from Chromium79, there were some problem with this solution when
the CPU architecture is same but the instruction length is
different(32 bit vs. 64 bit) with the target machine and the host
machine.

The proper solution would be using clang as a host machine compiler,
and this patch add configurations for that.

:Testing Performed:
Build succeed

:QA Notes:
Test-cases: N/A

:Issues Addressed:
[PLAT-131041] Add configurations to build qemux86

Change-Id: I0ba83921a2a52ff6cec074d9319950f66250536f
Originally-Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium84/+/296885
Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium87/+/301944
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Wanchang Ryu <wanchang.ryu@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Dmitry Ketov <dmitry.ketov@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Wanchang Ryu <wanchang.ryu@lge.com>
Tested-by: Test BTS <test.bts@lge.com>
Reviewed-on: http://gpro.lge.com/c/webosose/chromium87/+/307939
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>