[op][upsable][scroll] Fix scroll offset desynchronization
:Release Notes:
Avoid desynchronization of scroll offset if tree is updated but commits is not actually synchronized.

:Detailed Notes:
When scrolling in Photo&Video, that updates thumbnails while scroll is
still happening, we can observe sometimes the scroll lands in the
wrong place. This depends on the frame skipping and the code that
synchronizes the scroll offset between Impl and main CC code.

If the scroll offset is updated as part of the BeginMainFrame preparations,
but the frame is not actually emitted, the scroll is set to the
new value, and the SyncedScrollOffset value is updated, but the delta
is not fixed to refer to the new value. On next update, it will use
the wrong delta, and move further than expected. The solution
is making sure we reset the delta value at that point.

:Testing Performed:
Scroll photo&video on a folder with a lot of images (i.e. 10000)
and verify the landing position corresponds to the scroll done. It is
easier to reproduce remotely from web inspector, as it will synthesize
a wheel scroll that is bigger than regular remote wheel, so when the
issue happens the jump is bigger.

:QA Notes:
Test-cases: GT-32402

:Issues Addressed:
[PLAT-113352] PhotoVideo's item list is automatically jump to top or bottom sometimes after scrolling action end
[PLAT-132029] Port : [op][upsable][scroll] Fix scroll offset desynchronization

Change-Id: I23d948343c3cf973c114ce327d66962e78368d23
Originally-Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium84/+/294854
Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium87/+/303528
Tested-by: Jihye Hong <jh.hong@lge.com>
Tested-by: Test BTS <test.bts@lge.com>
Reviewed-by: Jihye Hong <jh.hong@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-on: http://gpro.lge.com/c/webosose/chromium87/+/308551
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>