[oe][ds][media][ls2] Match URIs array with URITypes
:Release Notes:
Match URIs array with URITypes

:Detailed Notes:
luna_service_uris array is indexed based on URIType, but there was
a mismatch between the sizes of each.
As some services are not available in OSE, use empty strings so that
there's no buffer overrun when accessing it based on any URIType.

:Testing Performed:
Verified that Media events from bluetooth headset are received from
MCS and the same are processed by the player. This ensures that the
service URI used to register with MCS is proper.

:QA Notes:
Test-cases: QO-1043

:Issues Addressed:
[PLAT-140649] [Port][Chromium87][oe][ds][media][ls2] Match URIs array
              with URITypes
[PLAT-117435] Match URIs array with URITypes

Change-Id: I1a5aa97005c21f46d7f4e005c2bdd3f9c586ce06
Originally-Reviewed-on: http://gpro.lge.com/c/290618
Reviewed-on: http://gpro.lge.com/c/webosose/chromium87/+/308080
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: <shridhar.kumbar@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: <shridhar.kumbar@lge.com>