[oe][n_upsable][app_shell][multi_display] Pass display ID from shell environment to app_shell
:Release Notes:
Pass DISPLAY_ID value from shell environment to app_shell binary

:Detailed Notes:
For launching app_shell instances in the right screen, we need to set
the display ID as a webOS window property (and set it as the current
display ID). In neva app runtime and webos webview we provide an API
that is used by WAM to set it.

For app_shell we need to read the environment variable DISPLAY_ID that
is passed from SAM to native applications.

:Testing Performed:
Launch enactbrowser in different displays and check it is visible in the
right screen.

:QA Notes:
Test-cases: QWA-2617

:Issues Addressed:
[PLAT-98000] Browser is launching in primary when launched from secondary
[PLAT-101335] Appshell type application should display itself based on DISPLAY_ID
[PLAT-139361] Port [oe][n_upsable][app_shell][multi_display] Pass display ID from shell environment to app_shell

Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium84/+/290836
Change-Id: I7c278df355ffd17778a7cf34d47887aafe1958b4
Reviewed-on: http://gpro.lge.com/c/webosose/chromium87/+/308339
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Vladislav Mukulov <vladislav.mukulov@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Test BTS <test.bts@lge.com>
Tested-by: Vladislav Mukulov <vladislav.mukulov@lge.com>