[op][upsable][wayland][touch] Use current time as timestamp for touch event
:Release Notes:
Used current time as timestamp (instead of the one sent by LSM) for
touch events to avoid possible invalidation (due to network delays).

:Detailed Notes:
The web content compares the timestamp of the thumbnails rendered
with the timestamp of click event fired to check the validation of
click event.

To generate click event from gesture tap, we use the timestamp sent from
LSM. But the timestamp is not valid sometimes because of some reasons
like a latency of touch events between LSM and Chromium, LSM sends whole
lot different timestamp from the platform time because of the load of
the whole system. In detail, timestamp in wl_touch is the elapsed time
since LSM process started up.

So this commit sets the timestamp to current time for TouchUp and
TouchDown event instead of using it from the Wayland server (LSM).

:Testing Performed:
1. Side-load a webpage containing scrollable and clickable items
   (e.g., https://looneypatterns.com)
2. Verify that right after items are scrolled on the page, clicking
   an item is properly handled (not lagging)

:QA Notes:
Test-cases: QWA-2347

:Issues Addressed:
[PLAT-99804]  Touch event not working for few seconds after
              scrolling in the app
[PLAT-130120] Port [op][upsable][wayland][touch] Use current
              time as timestamp for touch event

Originally-Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium84/+/294633
Change-Id: I0e75d4eac8e2cffb343e48641c0225693d3ed751
Reviewed-on: http://gpro.lge.com/c/webosose/chromium87/+/308670
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Test BTS <test.bts@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>