[oe][ds][ozone][input] Set the maximum input length to VKB
:Release Notes:
Implemented a way for setting the maximum input length to VKB

:Detailed Notes:
For VKB UI feature:
The Enter button will be disabled if the input field
entered the max allowed characters.
The app will set this max length of the input field and
the imemanger expected to receive from
inputMethodHost()->maxTextLength().
The Enter button will be enabled if the character length is
less than Max input length defined by the apps.

For the feature, in the patch, we get the maximum length
from HTML element and set it to VKB
via text model Wayland interface.

:Testing Performed:
RSPi4 webosose build.
Check Wayland log files and check that max input length
value is set to VKB via text model Wayland interface.

:QA Notes:
Test-cases: QSDK-5201

:Issues Addressed:
[PLAT-139370] Port [oe][ds][ozone][input] Set the maximum
              input length to VKB
[PLAT-89649]  Disable the Enter button after entering max
              characters

Change-Id: Ic41c41e658b3f55bb635dc7fa703f090a4f84f1f
Originally-Reviewed-on: http://gpro.lge.com/294883
Reviewed-on: http://gpro.lge.com/c/webosose/chromium87/+/308694
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Oleg Chaklyarov <oleg.chaklyarov@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Sergey Kipet <sergey.kipet@lge.com>
Tested-by: Test BTS <test.bts@lge.com>