[op][n_upsable][ozone][wayland] Implement virtual keyboard positioning
:Release Notes:
Provided capability of specifying virtual keyboard position and size
by defining a rectangle in surface local coordinates.

:Detailed Notes:
Due to the default virtual keyboard position and size doesn't always
fit user needs, functionality capable of specifying (on user's demand)
such traits has been requested.

The feature can be implemented by using the following already existing
API list of the text_model webOS Wayland extension:
* text_model_set_input_panel_rect
* text_model_reset_input_panel_rect

Thus, both origin location as well size of the virtual keyboard can be
specified by explicit defining the so called input panel rectangle.

So, the most tricky point with that regard is to provide that
capability, so the user could specify the rectangle from the web app
layer (either HTML or JS source code). To solve this, it was decided to
introduce a brand new HTML-attribute for the set of HTML input elements.
The attribute is named as 'inputelementcoords'. Below is a simple
example of specifying the input panel rectangle for a text input element
by using the HTML source code:

<input type="text" id="sn" name="sn" inputpanelcoords="100,273,1820,480">

When parsing the HTML-attribute the following constraints are checked:
* strictly four (4) coordinates are provided defining two (2) points -
  bottom left (x1,y1) and top right (x2,y2) corners of the specified
  rectangle
* the bottom left coordinates are followed by the top right coordinates
  when specified as "x1,y1,x2,y2"
* the specified coordinates lie inside the screen bounds

The changelist introduces the following:
* brand-new 'inputelementcoords' HTML-attribute
* handling of the introduced HTML-attribute by the Wayland backend

:Testing Performed:
1. Download the dedicated 'test.vkb.rectangle' test app
   onto the OSE board

2. Install the downloaded app:
   $ luna-send -i -f luna://com.webos.appInstallService/dev/install
     '{"id":"test.vkb.rectangle",
     "ipkUrl":"/home/root/test.vkb.rectangle_1.0.0_all.ipk",
     "subscribe":false}'

3. Launch the installed app:
   $ luna-send -n 1 -f luna://com.webos.applicationManager/launch
     '{"id":"test.vkb.rectangle"}'

4. Click either of the test app input elements

5. Ensure that the virtual keyboard behaves as expected (according
   to corresponding text label underneath each input element)

:QA Notes:
QWA-8800

:Issues Addressed:
[PLAT-126976] Export text_model::set_rect and text_model::reset_rect
              wayland calls through neva-extended PlatformWindow iface
[PLAT-141523] Port [op][n_upsable][ozone][wayland] Implement virtual
              keyboard positioning

Change-Id: I82ab3ea6e5de3c0b537e41c09a19faf20b2eb4d1
Originally-Reviewed-on: http://gpro.lge.com/307019
Reviewed-on: http://gpro.lge.com/c/webosose/chromium87/+/309385
Tested-by: Sergey Kipet <sergey.kipet@lge.com>
Tested-by: Test BTS <test.bts@lge.com>
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Vladislav Mukulov <vladislav.mukulov@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>