fixup! [op][n_upsable][web_security] Disable CORS/CORB checks for WAM applications
:Release Notes:
Fix iHeartRadio CORS issue with preflight redirect headers validation.

:Detailed Notes:
iHeartRadio failed on launching, even with allowCrossDomain, because
the error was happening in the preflight controller, and with the
type of error for unmatched headers for CORS.

This change adds the error to the ones to be rewritten in case
allowCrossDomain is set to true. And it also adds extra checking
for the variable in preflight controller.

:Testing Performed:
1. Load app with allowCrossDomain set to true
2. Side load iHeartRadio.
It should load.

1. Load app with allowCrossDomain set to true
2. Side load http://wpt.live/cors/redirect-preflight-2.htm
One of the tests SHOULD FAIL.

:QA Notes:
Test-cases: GT-46391

:Issues Addressed:
[PLAT-142451] [chromium87] Missing code to check CORS option for WAM

Change-Id: Ia173dd7e05675ccc09bd6a0c9269db6dfe21ae55
Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium87/+/310539
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-on: http://gpro.lge.com/c/webosose/chromium87/+/310594
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>